#5 Refactoring
With golangs zip implementation, it is not possible to add files to an existing archive.
The zip state handler now writes files in the work dir and creates an archive only at the end.

Furthermore, the state now contains a flag if the collectors are all done.
This flag is now used to determine reconciliation instead of the custom resource status.